refactor: move VerifierService to application
